Refactored utility functions to be more modular
