+++ api/functions/utils/cors.py
@@ -1,28 +1,5 @@
 from typing import List
 from flask import Response, abort
-from base64 import b64decode
-
-import firebase_admin
-from firebase_admin import credentials
-from firebase_admin import firestore
-
-
-def get_firestore_client():
-    project_id = "elpiscloud"
-
-    cred = credentials.ApplicationDefault()
-    default_app = firebase_admin.initialize_app(
-        cred,
-        {
-            "projectId": project_id,
-        },
-    )
-    return firestore.client()
-
-
-def decode_auth_header(user_info_header: str):
-    message_bytes = b64decode(user_info_header + "==")
-    return message_bytes.decode("ascii")
 
 
 def cors_preflight(request_methods: List[str]):
